Async/Awai data from modal

+++ assets/js/app.js
@@ -28,7 +28,6 @@ const listLocalStorage = "listShop";
 var idItem = "";
 let valorModal = 0;
 let arrItem = [];
-
 let dayWeekObj = {
   0: "Domingou no sofá😴? só que não! <br> Que tal iniciar a semana com uma bela lista de compras?",
   1: "Uma ótima Segunda-Feira pra começar aquele projetinho fitness 😅! <br> Pra não esquecer nada inicie uma lista de compras abaixo!",
@@ -56,14 +55,14 @@ function renderScreen() {
   if (!items || items.length === 0) {
     listBox.innerHTML = ""; //Se não houver nada salvo, não ocorre a renderização de elementos
     valueTotal.innerText = "R$ 0,00";
-    deleteAll.hidden = true; //se não há valores no array esconde os botões de delete
-    deleteChecked.hidden = true;
-    applyHidden(true);
+    //deleteAll.hidden = true; //se não há valores no array esconde os botões de delete
+    //deleteChecked.hidden = true;
+    applyHidden(true, "render");
     return;
   } else {
-    deleteAll.hidden = false; //se  há valores no array exibe os botões de delete
-    deleteChecked.hidden = false;
-    applyHidden(false);
+    //deleteAll.hidden = false; //se  há valores no array exibe os botões de delete
+    //deleteChecked.hidden = false;
+    applyHidden(false, "render");
   }
 
   //Laço para iteração e criação dos elementos HTMLs, identificando cada elemento com seu ID e prefixo definido pela natureza do elemento
@@ -86,75 +85,49 @@ function renderScreen() {
   listBox.innerHTML = htmlCode; //renderiza os componentes na tela <i class="fa-solid fa-circle-trash"></i>
 
   //-----------------Disparo dos eventos dinâmicos -----------------
-
   //Cria as funções em todos os 'buttons' para deletar items da lista
   var buttons = document.getElementsByClassName("delete"); // take all DOM elements buttons that have an 'remove'
-
   for (var i = 0; i < buttons.length; i++) {
     buttons[i].addEventListener("click", removeItem);
   }
 
   //Cria as funções em todos os elementos 'span' com a descrição do item, permitindo a edição após duploclick e fazendo a edição com o focusout
   var itemText = document.getElementsByClassName("lineThrough"); //  take all DOM elements span that have an 'Linethrough'
-
   for (var i = 0; i < itemText.length; i++) {
     itemText[i].addEventListener("focusout", editItem); //função focusout para alterar array/localstorage com base na descrição atualizada do elemento
     itemText[i].addEventListener("dblclick", contentEditable); //função para liberar a edição do elemento após duploclick
   }
-
   //Cria as funções para todos os checkboxs renderizados
   var check = document.getElementsByClassName("chk"); //
-
   for (var i = 0; i < buttons.length; i++) {
     check[i].addEventListener("click", boxChecked); //função para alterar o status checked do elemento no array/localstorage, assim como abrir modal para informar valores
   }
-
   //Por fim, é renderizado no footer o valor atualizado dos itens 'checkeds'
-
   sumTotal();
 }
 
 //Variavel funcional com as seguintes funcionalidades: Alteração do array/localstorage com atualização do 'checked',
 //inclusão de animações de movimento/alteração de cor do elemento, atualização do valor total da compra.
-var boxChecked = function () {
+var boxChecked = async function () {
   var parent = this.parentNode; //Obtém a referência do componente pai para inclusão das animações/alterações de background color
-
   var id = this.getAttribute("id"); // variável id criada para receber o id do componente 'checked'
-
   id = id.replace("chk-", ""); //Remove o prefixo determinado para obter o mesmo ID do array/localStorage
+  idItem = id;
 
-  arrItem = getList(); //Atualiza o array com base no localStorage
-
-  let result = arrItem.filter(function (el) {
-    return el.id == id; //filtra o array para obter um novo array apenas com o ID identificado
-  });
-
-  for (let element of result) {
-    let index = arrItem.indexOf(element);
-    arrItem[index].checked = this.checked; //altera o checked do valor do array com base no elemento marcado
-    !this.checked ? (arrItem[index].value = 0) : "";
-  }
-
-  localStorage.setItem(listLocalStorage, JSON.stringify(arrItem)); //altera o checked do valor do localStorage com base no elemento marcado
-  errModal.hidden = true;
+  this.checked ? await hiddenModal(false) : addValueArr(id,0,false);
 
   //verifica itens marcados e adiciona/remove classes para animação e cor no elemento, a depender do estado.
   if (this.checked) {
     parent.classList.add("checked", "animate__animated", "animate__pulse");
-    idItem = id;
-    modal.style.display = "block";
-    valueInputModal.focus();
+    spanModal.innerText = "";
+    idItem = "";
   } else {
     parent.classList.remove("checked", "animate__animated", "animate__pulse");
     parent.classList.add("noChecked");
-
     //realiza a atualização do valor total em tela - é reexecutada a função, pois para os efeitos de animação ocorrerem, a tela não pode ser renderizada novamente
-    sumTotal();
     //limpa o iD para o caso de um novo item ser marcado
     idItem = "";
   }
-  //
-  //
 };
 
 //Função chamada para permitir a edição do item com o duploclick no elemento
@@ -180,8 +153,7 @@ var editItem = function () {
     arrItem[index].text = text; //atualiza o array com informações do item editado
   }
 
-  localStorage.setItem(listLocalStorage, JSON.stringify(arrItem)); //Atualiza o localstorage a partir do array editado
-  renderScreen(); // Renderiza a tela com as novas informações
+  updateLocalStorage()
 };
 
 //Função para remover item do Array/Localstorage
@@ -200,8 +172,7 @@ var removeItem = function () {
     arrItem.splice(index, 1);
   }
 
-  localStorage.setItem(listLocalStorage, JSON.stringify(arrItem)); //Atualiza a lista no localStorage após remoção do item
-  renderScreen(); // Renderiza a tela
+  updateLocalStorage()
 };
 
 // Função para gerar novo ID e adicionar item ao array
@@ -223,21 +194,14 @@ const getList = function () {
 //Renderizar itens na tela no load do formulário
 renderScreen();
 
-//Captura o click no modal e realiza a atualização do valor do item no array/LocalStorage
-valueBtnModal.addEventListener("click", () => {
-  if (valueInputModal.value) {
-    valorModal = parseFloat(valueInputModal.value);
-    addValueArr(idItem, valorModal);
-    valueInputModal.value = "";
-    modal.style.display = "none";
-    errModal.hidden = true;
-  } else {
-    errModal.hidden = false;
-  }
-});
+
+function updateLocalStorage(){
+  localStorage.setItem(listLocalStorage, JSON.stringify(arrItem)); //
+  renderScreen(); // Render Screen
+}
 
 // Variavel funcional para adicionar valor no  array/localstorage
-var addValueArr = (idItem, valor) => {
+var addValueArr = (idItem, valor, checkedValue) => {
   let result = arrItem.filter(function (el) {
     return el.id == idItem;
   });
@@ -245,10 +209,10 @@ var addValueArr = (idItem, valor) => {
   for (let element of result) {
     let index = arrItem.indexOf(element);
     arrItem[index].value = valor;
+    arrItem[index].checked = checkedValue;
+    !checkedValue ? (arrItem[index].value = 0) : "";
   }
-
-  localStorage.setItem(listLocalStorage, JSON.stringify(arrItem)); //
-  renderScreen(); // Render Screen
+  updateLocalStorage()
 };
 
 //Captura evento submit no formulário para adição de novos itens
@@ -264,19 +228,22 @@ formSubmit.addEventListener("submit", (e) => {
   }
 
   addItem(inputItem.value.trim(), 0, 0); //função para adição de novos itens
-  localStorage.setItem(listLocalStorage, JSON.stringify(arrItem)); //atualiza localstorage com informações do array
+  updateLocalStorage()
   inputItem.value = ""; //limpa o campo após salvar dados
   inputItem.focus(); //atualiza o foco no campo input para inserção de novos itens
-  renderScreen(); //renderiza a tela com itens adicionados
 });
 
 // Função para deletar todos os itens do Array/Objeto atualizando a tela
 function deleteAllItems() {
+  const validateDelete = confirm('Você realmente deseja deletar toda a lista de compras?')
+
+  if (validateDelete){
   localStorage.clear();
   arrItem = [];
   renderScreen();
   deleteAll.hidden = true;
   deleteChecked.hidden = true;
+  }
 }
 
 //função para deletar apenas itens checkados do array/localstorage ao fim renderizando a tela;
@@ -292,8 +259,7 @@ function deleteCheckeds() {
     arrItem.splice(index, 1);
   }
 
-  localStorage.setItem(listLocalStorage, JSON.stringify(arrItem)); //
-  renderScreen(); // Render Screen
+  updateLocalStorage()
 }
 
 //Função para soma do total com base no array atualizado pelo localstorage
@@ -320,8 +286,45 @@ deleteChecked.addEventListener("click", () => deleteCheckeds());
 
 btnMain.addEventListener("click", () => applyHidden(false));
 
-function applyHidden(value) {
+function applyHidden(value, render) {
   formSubmit.hidden = value;
   divFooter.hidden = value;
   divEmptyList.hidden = !value;
+  //ternário para exibir/remover os botões apenas na função reenderscreen quando houver itens no array/localstorage
+  //assim, quando não há itens cadastrados na tela, apenas os botões deleteall/deleteChecked não são exibidos
+  render ? ((deleteAll.hidden = value), (deleteChecked.hidden = value)) : "";
+}
+
+async function hiddenModal(value) {
+  return new Promise((resolve) => {
+    let stop = false;
+    errModal.hidden = !value;
+    valueInputModal.focus();
+    modal.style.display = "block";
+
+    valueBtnModal.addEventListener("click", () => {
+      if (valueInputModal.value) {
+        valorModal = parseFloat(valueInputModal.value);
+        addValueArr(idItem, valorModal, true);
+        valueInputModal.value = "";
+        modal.style.display = "none";
+        errModal.hidden = true;
+        stop = true;
+      } else {
+        errModal.hidden = false;
+        stop = true;
+      }
+
+      loop();
+
+      function loop() {
+        if (stop === true) {
+          resolve();
+          return;
+        }
+
+        setTimeout(loop, 1000);
+      }
+    });
+  });
 }
